Фикс геймификации и добавление мелких фич для удобства фронтенда

+++ src/main/java/com/example/online_bank/domain/event/UpdateBonusAccountEvent.java
@@ -1,7 +1,9 @@
 package com.example.online_bank.domain.event;
 
+import java.math.BigDecimal;
+
 public record UpdateBonusAccountEvent(
-        Integer points,
+        BigDecimal points,
         String accountNumber
 ) {
 }
